Check if /var/lib/ceph/nss exists before creation
mkdir() throws an exception if the directory already exists, so we should check
for it's existance before creation as hooks should be idempotent.

This occurred in practice when the install hook had to be re-run, following a
previous failure due to juju API timeout.  It would also occur if the keystone
relation was re-built, so that case is also fixed.

Change-Id: I4052cda5bb20f76ab592ed7817bdc1e5b5b2138d
Closes-Bug: #1563667