Add support for cephx pool grouping and permissions
Sync charmhelpers and add configuration option to allow access
to ceph pools to be limited based on grouping.

The radosgw will only require access to pools in the 'objects'
group - which probably won't be shared with other services,
but does ensure that compromise on a radosgw unit does not
allow access to volumes and instances on the same underlying
ceph backend.

Added creation of .rgw.root pool to broker requests to ensure
that all pools are added to the 'objects' group, avoiding the
need for rw permissionson the radosgw.gateway key.

Change-Id: I88953b44e9994c1d0235f617faaeb892b7abed9c
Partial-Bug: 1424771