Switch charm runtime to python3
Update charm to execute hooks and actions under Python 3; this
includes dealing with upgrades (by switching the upgrade-charm
hook to be a bash script which installs the required charm runtime
dependencies).

This commit also drops code from ceph.py which was used across
other ceph charms in the past; only the functions required for this
charm have been retained.

Change-Id: I5e222d907bfa34ffacad16c51abd1278d7d82f56