Align context and interface name for mon relation
Ensure that the name of the local endpoint and the context are
aligned - 'ceph-radosgw' is the type of the interface and not
the name of the local endpoint which is 'mon'.

This misnaming causes the charm to sit in blocked state for an
extended period of time until the relation data bag is complete,
skipping the intermediate waiting/incomplete state.

Change-Id: I9d3fab6020817be23f26d410aa59878287dbf5b0
Closes-Bug: 1885144