Update README for Ceph EC pools
This updates the README for erasure coded
Ceph pools for the case of Ceph-based Object
storage.

The new text should be as similar as possible
for all the charms that support configuration
options for EC pools. See the below review for
the first of these charms whose README has been
updated.

https://review.opendev.org/#/c/749824/

Remove section on RGW multisite replication as
it is already in the CDG. The charm now points
there.

Add basic README template sections (Actions,
Bugs, Configuration). Standardise Overview,
Deployment, and Network spaces sections. Improve
Access and Keystone integration sections.

Change-Id: I3839c1018b9bdf0d6712d3fb2e9f95b633591615