Remove reference to ntp package
This is motivated by the fact that NTP will be managed by
operators through the NTP charm. In the other Ceph charms,
NTP / Chrony are removed by the charm if the charm is in a
container because the ceph-base package will install it;
however, this is unneceessary for the radosgw package
because it doesn't depend on the ceph-common package.

Change-Id: I53ab9ece631bbaeed6e3cebba4ffbbcb1e32193d
Closes-Bug: #1915351